LayoutTests/imported/w3c/web-platform-tests/content-security-policy/embedded-enforcement/subsumption_algorithm-unsafe_eval-expected.txt

Blocked access to external URL http://xn--n8j6ds53lwwkrqhv28a.localhost:8800/content-security-policy/embedded-enforcement/support/echo-policy-multiple.py?policy=style-src+http%3A%2F%2Fexample1.com%2Ffoo%2Fbar.html+%27unsafe-eval%27&id=%27unsafe-eval%27+is+properly+subsumed.
Blocked access to external URL http://xn--n8j6ds53lwwkrqhv28a.localhost:8800/content-security-policy/embedded-enforcement/support/echo-policy-multiple.py?policy=style-src+http%3A%2F%2Fexample1.com%2Ffoo%2F+%27unsafe-inline%27&id=No+other+keyword+has+the+same+effect+as+%27unsafe-eval%27.
Blocked access to external URL http://xn--n8j6ds53lwwkrqhv28a.localhost:8800/content-security-policy/embedded-enforcement/support/echo-policy-multiple.py?policy=style-src+http%3A%2F%2Fexample1.com%2Ffoo%2F+%27unsafe-inline%27+%27unsafe-eval%27&id=Other+expressions+have+to+be+subsumed.
Blocked access to external URL http://xn--n8j6ds53lwwkrqhv28a.localhost:8800/content-security-policy/embedded-enforcement/support/echo-policy-multiple.py?policy=style-src+http%3A%2F%2Fexample1.com%2Ffoo%2F+%27unsafe-hashed-attributes%27+%27unsafe-eval%27&policy2=style-src+http%3A%2F%2Fexample1.com%2Ffoo%2F+%27self%27+%27unsafe-eval%27&id=Effective+policy+is+properly+found.
Blocked access to external URL http://xn--n8j6ds53lwwkrqhv28a.localhost:8800/content-security-policy/embedded-enforcement/support/echo-policy-multiple.py?policy=style-src+http%3A%2F%2Fexample1.com%2Ffoo%2F+%27self%27++%27unsafe-eval%27&id=Required+csp+must+allow+%27unsafe-eval%27.
Blocked access to external URL http://xn--n8j6ds53lwwkrqhv28a.localhost:8800/content-security-policy/embedded-enforcement/support/echo-policy-multiple.py?policy=style-src+%27unsafe-hashed-attributes%27+%27unsafe-eval%27&policy2=style-src+%27unsafe-eval%27+%27unsafe-inline%27&id=Effective+policy+is+properly+found+where+%27unsafe-eval%27+is+not+subsumed.
Blocked access to external URL http://xn--n8j6ds53lwwkrqhv28a.localhost:8800/content-security-policy/embedded-enforcement/support/echo-policy-multiple.py?policy=style-src+%27unsafe-hashed-attributes%27+%27self%27&policy2=style-src+%27unsafe-eval%27+%27self%27&id=Effective+policy+is+properly+found+where+%27unsafe-eval%27+is+not+part+of+it.


Harness Error (TIMEOUT), message = null

TIMEOUT 'unsafe-eval' is properly subsumed. Test timed out
FAIL No other keyword has the same effect as 'unsafe-eval'. assert_throws_dom: function "() => {
        var x = i.contentWindow.location.href;
      }" did not throw
FAIL Other expressions have to be subsumed. assert_throws_dom: function "() => {
        var x = i.contentWindow.location.href;
      }" did not throw
TIMEOUT Effective policy is properly found. Test timed out
FAIL Required csp must allow 'unsafe-eval'. assert_throws_dom: function "() => {
        var x = i.contentWindow.location.href;
      }" did not throw
FAIL Effective policy is properly found where 'unsafe-eval' is not subsumed. assert_throws_dom: function "() => {
        var x = i.contentWindow.location.href;
      }" did not throw
TIMEOUT Effective policy is properly found where 'unsafe-eval' is not part of it. Test timed out
